+++ android/expoview/src/main/java/host/exp/exponent/kernel/services/linking/LinkingKernelService.kt
@@ -2,14 +2,12 @@
 package host.exp.exponent.kernel.services.linking
 
 import android.net.Uri
-import host.exp.exponent.Constants
 import host.exp.exponent.kernel.KernelConstants.ExperienceOptions
 import host.exp.exponent.kernel.KernelProvider
 
 class LinkingKernelService {
   fun openURI(uri: Uri) {
-    val manifestUrl =
-      if (Constants.isStandaloneApp()) Constants.INITIAL_URL.toString() else uri.toString()
+    val manifestUrl = uri.toString()
     KernelProvider.instance
       .openExperience(ExperienceOptions(manifestUrl, uri.toString(), null))
   }
@@ -17,9 +15,6 @@ class LinkingKernelService {
   fun canOpenURI(uri: Uri): Boolean {
     val scheme = uri.scheme
     if (scheme == "exp" || scheme == "exps") {
-      return true
-    }
-    if (Constants.SHELL_APP_SCHEME != null && Constants.SHELL_APP_SCHEME == scheme) {
       return true
     }
     val host = uri.host