+++ android/versioned-abis/expoview-abi48_0_0/src/main/java/abi48_0_0/expo/modules/kotlin/allocators/ObjectConstructor.kt
@@ -1,5 +1,0 @@
-package abi48_0_0.expo.modules.kotlin.allocators
-
-fun interface ObjectConstructor<T> {
-  fun construct(): T
-}