+++ android/versioned-abis/expoview-abi48_0_0/src/main/java/abi48_0_0/expo/modules/kotlin/views/CallbacksDefinition.kt
@@ -1,3 +1,0 @@
-package abi48_0_0.expo.modules.kotlin.views
-
-class CallbacksDefinition(val names: Array<out String>)