+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -97,12 +97,14 @@ export function withExtendedResolver(
     platforms,
     isTsconfigPathsEnabled,
     isFastResolverEnabled,
+    isExporting,
   }: {
     projectRoot: string;
     tsconfig: TsConfigPaths | null;
     platforms: string[];
     isTsconfigPathsEnabled?: boolean;
     isFastResolverEnabled?: boolean;
+    isExporting?: boolean;
   }
 ) {
   if (isFastResolverEnabled) {
@@ -165,36 +167,40 @@ export function withExtendedResolver(
       })
     : null;
 
-  if (isTsconfigPathsEnabled && isInteractive()) {
-    // TODO: We should track all the files that used imports and invalidate them
-    // currently the user will need to save all the files that use imports to
-    // use the new aliases.
-    const configWatcher = new FileNotifier(projectRoot, ['./tsconfig.json', './jsconfig.json']);
-    configWatcher.startObserving(() => {
-      debug('Reloading tsconfig.json');
-      loadTsConfigPathsAsync(projectRoot).then((tsConfigPaths) => {
-        if (tsConfigPaths?.paths && !!Object.keys(tsConfigPaths.paths).length) {
-          debug('Enabling tsconfig.json paths support');
-          tsConfigResolve = resolveWithTsConfigPaths.bind(resolveWithTsConfigPaths, {
-            paths: tsConfigPaths.paths ?? {},
-            baseUrl: tsConfigPaths.baseUrl,
-          });
-        } else {
-          debug('Disabling tsconfig.json paths support');
-          tsConfigResolve = null;
-        }
+  if (!isExporting && isInteractive()) {
+    if (isTsconfigPathsEnabled) {
+      // TODO: We should track all the files that used imports and invalidate them
+      // currently the user will need to save all the files that use imports to
+      // use the new aliases.
+      const configWatcher = new FileNotifier(projectRoot, ['./tsconfig.json', './jsconfig.json']);
+      configWatcher.startObserving(() => {
+        debug('Reloading tsconfig.json');
+        loadTsConfigPathsAsync(projectRoot).then((tsConfigPaths) => {
+          if (tsConfigPaths?.paths && !!Object.keys(tsConfigPaths.paths).length) {
+            debug('Enabling tsconfig.json paths support');
+            tsConfigResolve = resolveWithTsConfigPaths.bind(resolveWithTsConfigPaths, {
+              paths: tsConfigPaths.paths ?? {},
+              baseUrl: tsConfigPaths.baseUrl,
+            });
+          } else {
+            debug('Disabling tsconfig.json paths support');
+            tsConfigResolve = null;
+          }
+        });
       });
-    });
 
-    // TODO: This probably prevents the process from exiting.
-    installExitHooks(() => {
-      configWatcher.stopObserving();
-    });
-  } else {
-    debug('Skipping tsconfig.json paths support');
+      // TODO: This probably prevents the process from exiting.
+      installExitHooks(() => {
+        configWatcher.stopObserving();
+      });
+    } else {
+      debug('Skipping tsconfig.json paths support');
+    }
   }
 
   let nodejsSourceExtensions: string[] | null = null;
+
+  const shimsFolder = path.join(require.resolve('@expo/cli/package.json'), '..', 'static/shims');
 
   return withMetroResolvers(config, projectRoot, [
     // Add a resolver to alias the web asset resolver.
@@ -372,6 +378,18 @@ export function withExtendedResolver(
         ) {
           // @ts-expect-error: `readonly` for some reason.
           result.filePath = reactNativeWebAppContainer;
+        } else if (platform === 'web' && result.filePath.includes('node_modules')) {
+          // Replace with static shims
+
+          const normalName = normalizeSlashes(result.filePath)
+            // Drop everything up until the `node_modules` folder.
+            .replace(/.*node_modules\//, '');
+
+          const shimPath = path.join(shimsFolder, normalName);
+          if (fs.existsSync(shimPath)) {
+            // @ts-expect-error: `readonly` for some reason.
+            result.filePath = shimPath;
+          }
         }
       }
       return result;
@@ -419,6 +437,7 @@ export async function withMetroMultiPlatformAsync(
     webOutput,
     routerDirectory,
     isFastResolverEnabled,
+    isExporting,
   }: {
     config: ConfigT;
     isTsconfigPathsEnabled: boolean;
@@ -426,6 +445,7 @@ export async function withMetroMultiPlatformAsync(
     webOutput?: 'single' | 'static' | 'server';
     routerDirectory: string;
     isFastResolverEnabled?: boolean;
+    isExporting?: boolean;
   }
 ) {
   // Auto pick app entry for router.
@@ -464,6 +484,7 @@ export async function withMetroMultiPlatformAsync(
     tsconfig,
     isTsconfigPathsEnabled,
     isFastResolverEnabled,
+    isExporting,
   });
 }
 
@@ -475,12 +496,14 @@ function withMetroMultiPlatform(
     isTsconfigPathsEnabled,
     tsconfig,
     isFastResolverEnabled,
+    isExporting,
   }: {
     config: ConfigT;
     isTsconfigPathsEnabled: boolean;
     platformBundlers: PlatformBundlers;
     tsconfig: TsConfigPaths | null;
     isFastResolverEnabled?: boolean;
+    isExporting?: boolean;
   }
 ) {
   let expoConfigPlatforms = Object.entries(platformBundlers)
@@ -501,6 +524,7 @@ function withMetroMultiPlatform(
   return withExtendedResolver(config, {
     projectRoot,
     tsconfig,
+    isExporting,
     isTsconfigPathsEnabled,
     platforms: expoConfigPlatforms,
     isFastResolverEnabled,