+++ packages/expo-build-properties/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### 🎉 New features
 
-- Add `withAndroidQueries` to allow adding and modifying the `<queries>` tag in the App Manifest.
+- Add `withAndroidQueries` to allow adding and modifying the `<queries>` tag in the App Manifest. ([#24619](https://github.com/expo/expo/pull/24619) by [@alanjhughes](https://github.com/alanjhughes))
 
 ### 🐛 Bug fixes